refactor: Quest Update 로직 분리
