+++ src/main/java/com/groom/orbit/goal/controller/QuestCommandController.java
@@ -14,6 +14,7 @@ import com.groom.orbit.common.annotation.AuthMember;
 import com.groom.orbit.common.dto.CommonSuccessDto;
 import com.groom.orbit.common.dto.ResponseDto;
 import com.groom.orbit.goal.app.QuestCommandService;
+import com.groom.orbit.goal.app.QuestUpdateService;
 import com.groom.orbit.goal.app.dto.CreateQuestRequestDto;
 import com.groom.orbit.goal.app.dto.UpdateQuestRequestDto;
 import com.groom.orbit.goal.app.dto.UpdateQuestSequenceRequestDto;
@@ -26,6 +27,7 @@ import lombok.RequiredArgsConstructor;
 public class QuestCommandController {
 
   private final QuestCommandService questCommandService;
+  private final QuestUpdateService questUpdateService;
 
   @PutMapping
   public ResponseDto<CommonSuccessDto> createQuest(
@@ -46,12 +48,12 @@ public class QuestCommandController {
       @AuthMember Long memberId,
       @PathVariable("quest_id") Long questId,
       @RequestBody UpdateQuestRequestDto dto) {
-    return ResponseDto.ok(questCommandService.updateQuest(memberId, questId, dto));
+    return ResponseDto.ok(questUpdateService.updateQuest(memberId, questId, dto));
   }
 
   @PatchMapping
   public ResponseDto<CommonSuccessDto> updateQuestsSequence(
       @AuthMember Long memberId, @RequestBody List<UpdateQuestSequenceRequestDto> dtos) {
-    return ResponseDto.ok(questCommandService.updateQuestSequence(memberId, dtos));
+    return ResponseDto.ok(questUpdateService.updateQuestSequence(memberId, dtos));
   }
 }
